pyproject.toml
[project]
name = "livekit-transient"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "livekit>=1.0.13",
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-deepgram>=1.2.14",
    "livekit-plugins-elevenlabs>=1.2.14",
    "livekit-plugins-openai>=1.2.14",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.37.0",
]
